Removed classes added to spi to reimplement services that uses sling-common-crypto
