+++ src/main/java/org/apache/sling/auth/oauth_client/impl/JcrUserHomeOAuthTokenStore.java
@@ -30,11 +30,6 @@ import javax.jcr.Value;
 import org.apache.jackrabbit.api.security.user.User;
 import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.auth.oauth_client.ClientConnection;
-import org.apache.sling.auth.oauth_client.spi.OAuthException;
-import org.apache.sling.auth.oauth_client.spi.OAuthToken;
-import org.apache.sling.auth.oauth_client.spi.OAuthTokenStore;
-import org.apache.sling.auth.oauth_client.spi.OAuthTokens;
-import org.apache.sling.auth.oauth_client.spi.TokenState;
 import org.jetbrains.annotations.NotNull;
 import org.apache.sling.commons.crypto.CryptoService;
 import org.osgi.service.component.annotations.Activate;
@@ -56,7 +51,7 @@ public class JcrUserHomeOAuthTokenStore implements OAuthTokenStore {
     private CryptoService cryptoService;
     
     @Activate
-    public JcrUserHomeOAuthTokenStore(@Reference(target = "(names=sling-oauth)") CryptoService cryptoService) {
+    public JcrUserHomeOAuthTokenStore(@Reference CryptoService cryptoService) {
         this.cryptoService = cryptoService;
     }
