+++ src/main/java/org/apache/sling/auth/oauth_client/impl/OAuthException.java
@@ -14,10 +14,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.sling.auth.oauth_client.spi;
+package org.apache.sling.auth.oauth_client.impl;
 
-import org.jetbrains.annotations.Nullable;
+public class OAuthException extends RuntimeException {
 
-public record OAuthTokens(@Nullable String accessToken, long expiresAt, @Nullable String refreshToken) {
+    private static final long serialVersionUID = 1L;
+
+    public OAuthException() {
+        super();
+    }
+
+    public OAuthException(String message, Throwable cause) {
+        super(message, cause);
+    }
+
+    public OAuthException(String message) {
+        super(message);
+    }
+
+    public OAuthException(Throwable cause) {
+        super(cause);
+    }
 
 }
