+++ src/main/java/org/apache/sling/auth/oauth_client/impl/OAuthState.java
@@ -14,26 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.sling.auth.oauth_client.spi;
+package org.apache.sling.auth.oauth_client.impl;
 
-public class OAuthException extends RuntimeException {
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
-    private static final long serialVersionUID = 1L;
-
-    public OAuthException() {
-        super();
-    }
-
-    public OAuthException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public OAuthException(String message) {
-        super(message);
-    }
-
-    public OAuthException(Throwable cause) {
-        super(cause);
-    }
-
-}
+public record OAuthState(@NotNull String perRequestKey, @NotNull String connectionName, @Nullable String redirect) {}
