+++ src/main/java/org/apache/sling/auth/oauth_client/impl/OAuthTokenRefresher.java
@@ -18,8 +18,6 @@ package org.apache.sling.auth.oauth_client.impl;
 
 import org.apache.sling.auth.oauth_client.ClientConnection;
 
-import org.apache.sling.auth.oauth_client.spi.OAuthException;
-import org.apache.sling.auth.oauth_client.spi.OAuthTokens;
 import org.jetbrains.annotations.NotNull;
 
 public interface OAuthTokenRefresher {
