+++ src/main/java/org/apache/sling/auth/oauth_client/impl/OidcProviderMetadataRegistry.java
@@ -21,7 +21,6 @@ import java.net.URI;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
-import org.apache.sling.auth.oauth_client.spi.OAuthException;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.osgi.service.component.annotations.Component;
