+++ src/main/java/org/apache/sling/auth/oauth_client/impl/OidcTokens.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.sling.auth.oauth_client.spi;
+package org.apache.sling.auth.oauth_client.impl;
 
 import org.jetbrains.annotations.Nullable;
