+++ src/main/java/org/apache/sling/auth/oauth_client/impl/TokenAccessImpl.java
@@ -23,10 +23,6 @@ import org.apache.sling.api.resource.ResourceResolver;
 import org.apache.sling.auth.oauth_client.ClientConnection;
 import org.apache.sling.auth.oauth_client.OAuthTokenAccess;
 import org.apache.sling.auth.oauth_client.OAuthTokenResponse;
-import org.apache.sling.auth.oauth_client.spi.OAuthToken;
-import org.apache.sling.auth.oauth_client.spi.OAuthTokenStore;
-import org.apache.sling.auth.oauth_client.spi.OAuthTokens;
-import org.apache.sling.auth.oauth_client.spi.TokenState;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
